SuitSorter API
An API for developers to allow them to sort suits optimally. This implementation lays on top of the mod itself, allowing it to be safe. In the scenario of a SuitSorter/plugin being broken (for whatever reason), it will not effect the main mod itself. The API supports both loading by config or swapping at runtime (even in the middle of a round); however, the runtime feature is up to a mod to implement and use.